1st successful int. test w/ Read All, Plan & Execute

+++ inc/flexsea_system.h
@@ -51,7 +51,7 @@
 //****************************************************************************
 
 void init_flexsea_payload_ptr(void);
-uint32_t tx_cmd(uint8_t *payloadData, uint8_t cmdCode, uint8_t cmd_type, \
+uint16_t tx_cmd(uint8_t *payloadData, uint8_t cmdCode, uint8_t cmd_type, \
 				uint32_t len, uint8_t receiver, uint8_t *buf);
 __attribute__((weak)) void init_flexsea_payload_ptr_user(void);
